+++ app/api/collectionchildren/[slug]/route.tsx
@@ -1,5 +1,5 @@
 import { NextRequest } from "next/server";
-import { fetchApi } from "@/src/utils/fetchApi";
+import { fetchApi } from "@/src/utils/fetchApi/fetchApi";
 import { CARDS_PER_PAGE } from "@/src/config/constants";
 
 const oneMonth = 60 * 60 * 24 * 30;
