+++ app/api/lanes/route.tsx
@@ -1,5 +1,5 @@
 import { NextResponse } from "next/server";
-import lanesData from "../../src/data/lanes";
+import lanesData from "../../src/data/lanesData";
 
 export const GET = async () => {
   return NextResponse.json(lanesData, { status: 200 });
