+++ app/divisions/[slug]/page.tsx
@@ -2,7 +2,7 @@ import { Metadata } from "next";
 import React from "react";
 import DivisionPage from "../../src/components/pages/divisionPage/divisionPage";
 import { slugToString } from "../../src/utils/utils";
-import { RepoApi } from "../../src/utils/apiClients";
+import { RepoApi } from "../../src/utils/apiClients/apiClients";
 import { Suspense } from "react";
 import { redirect } from "next/navigation";
