+++ app/src/components/aboutPageElements/aboutPageElements.tsx
