+++ app/src/components/pages/aboutPage/aboutPage.tsx
@@ -1,6 +1,6 @@
 "use client";
 import { Box } from "@nypl/design-system-react-components";
-import aboutPageElements from "../../../data/aboutPageElements";
+import aboutPageElements from "../../aboutPageElements/aboutPageElements";
 import PageLayout from "../../pageLayout/pageLayout";
 import { createAdobeAnalyticsPageName } from "@/src/utils/utils";
