+++ app/src/components/search/filters/rightsFilter.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen, fireEvent } from "@testing-library/react";
 import RightsFilter from "./rightsFilter";
 import { useRouter, usePathname } from "next/navigation";
-import { SearchManager } from "@/src/utils/searchManager";
+import { SearchManager } from "@/src/utils/searchManager/searchManager";
 
 jest.mock("next/navigation", () => ({
   useRouter: jest.fn(),
