+++ app/src/components/search/filters/selectFilterModal.test.tsx
@@ -5,7 +5,7 @@ import {
   DEFAULT_SEARCH_SORT,
   DEFAULT_SEARCH_TERM,
 } from "@/src/config/constants";
-import { GeneralSearchManager } from "@/src/utils/searchManager";
+import { GeneralSearchManager } from "@/src/utils/searchManager/searchManager";
 import { AvailableFilter } from "@/src/types/AvailableFilterType";
 
 jest.mock("@chakra-ui/react", () => {
