--- app/src/types/CollectionSearchParams.ts
@@ -1,11 +1,0 @@
-import { AvailableFilterOption } from "./AvailableFilterType";
-
-export type CollectionSearchParamsType = {
-  q: string;
-  sort: string;
-  filters: string;
-  page: number;
-  availableFilters?: Record<string, AvailableFilterOption[]>;
-};
-
-export default CollectionSearchParamsType;
